requirements.txt
Appium-Python-Client==0.30
AxmlParserPY==0.1
numpy==1.12.0
opencv-python==3.4.3.18
selenium==3.141.0
urllib3==1.24.1
